requirements.txt
deepface
streamlit
opencv-python
numpy
scikit-learn
duckduckgo-search
pillow
facenet-pytorch